Updates Formula: ufraw
UFRaw requires GTK to build the graphical portion. Using the
`--without-gtk` option to configure will onluy build a batch command
line. This is the default behavior for this formula.

This commit adds an option to the formula so the user can choose to
include the GTK dependancy and build the graphical front end for UFRaw.